completed
Had to modify the test at the beginning of Part 2, as attempting to directly compare was resulting in it flagging that my numbers did not exactly match up. However, I do not see anything different I could have done to get the matching numbers, and suspect it is an issue of jpg artifacting causing the misalignment.